KernelBackend/requirements.txt
django
psycopg2
djangorestframework
django-cors-headers
openpyxl
